change int to long in sales

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -3,15 +3,16 @@ package ru.netology.stats;
 public class StatsService {
 
     // Сумма всех продаж
-    public int sumSales(int[] sales){
+    public long sumSales(int[] sales) {
         int sum = 0;
-        for (int sale : sales){
+        for (int sale : sales) {
             sum += sale;
         }
         return sum;
     }
+
     //  Минимальная сумма продаж
-    public int minSales(int[] sales) {
+    public long minSales(int[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -21,8 +22,9 @@ public class StatsService {
 
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
+
     // Максимальная сумма продаж
-    public int maxSales(int[] sales) {
+    public long maxSales(int[] sales) {
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) {
@@ -31,28 +33,31 @@ public class StatsService {
         }
         return maxMonth + 1;
     }
+
     // Средняя сумма продаж
-    public int averageSales(int[] sales) {
-       return sumSales(sales) / sales.length;
+    public long averageSales(int[] sales) {
+        return sumSales(sales) / sales.length;
     }
+
     // продажами ниже среднего
-    public int belowAverageSales(int[] sales){
-        int average = averageSales(sales);
+    public long belowAverageSales(int[] sales) {
+        long average = averageSales(sales);
         int count = 0;
-        for (int sale : sales){
-            if (sale > average){
-                count ++;
+        for (int sale : sales) {
+            if (sale > average) {
+                count++;
             }
         }
         return count;
     }
+
     // продажами выше среднего
-    public int aboveAverageSales(int[] sales){
-        int average = averageSales(sales);
+    public long aboveAverageSales(int[] sales) {
+        long average = averageSales(sales);
         int count = 0;
-        for (int sale : sales){
-            if (sale < average){
-                count ++;
+        for (int sale : sales) {
+            if (sale < average) {
+                count++;
             }
         }
         return count;
